requirements.txt
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
scikit-learn==1.4.1.post1
pandas==2.2.1
spacy==3.7.4
nltk==3.8.1
opencv-python==4.11.0.86
pytesseract==0.3.10
Flask==3.0.2
Flask-SQLAlchemy==3.1.1
scikit-learn==1.4.1.post1
pandas==2.2.1
spacy==3.7.4
nltk==3.8.1
opencv-python==4.11.0.86
pytesseract==0.3.10
Werkzeug==3.0.1
pdfplumber==0.11.0
SpeechRecognition==3.10.3
pydub==0.25.1
requests==2.31.0
numpy==1.26.4
python-dotenv==1.0.0
transformers
torch
requests
python-whois
tldextract
